Fix serial channel timeout during stop
If stop is called at the same time of the process.kill() for timer, the
system will hang for process being NoneType.

+++ ltp/channel/shell.py
@@ -107,6 +107,9 @@ class ShellChannel(Channel):
 
             if t_secs >= 0:
                 def _threaded():
+                    if self._process or self._stop:
+                        return
+
                     self._logger.info(
                         "Command timed out after %d seconds", timeout)
                     self._process.kill()
